Refactoring the producer to work well with configuration file
