+++ kafka-producer-example/src/main/java/com/javatechie/constants/KafkaTopicNames.java
@@ -1,0 +1,7 @@
+package com.javatechie.constants;
+
+public class KafkaTopicNames {
+    public static final String TOPIC_NAME_CUSTOMER_DATA = "javatechie-customer-data";
+    public static final String TOPIC_NAME_TEXT_DATA = "javatechie-text-data";
+
+}
