+++ kafka-producer-example/src/main/java/com/javatechie/controller/EventController.java
@@ -1,7 +1,8 @@
 package com.javatechie.controller;
 
 import com.javatechie.dto.Customer;
-import com.javatechie.service.KafkaMessagePublisher;
+import com.javatechie.service.KafkaEventMessagePublisher;
+import com.javatechie.service.KafkaPlainTextMessagePublisher;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,13 +13,16 @@ import org.springframework.web.bind.annotation.*;
 public class EventController {
 
     @Autowired
-    private KafkaMessagePublisher publisher;
+    private KafkaPlainTextMessagePublisher plainTextMessagePublisher;
+
+    @Autowired
+    private KafkaEventMessagePublisher eventMessagePublisher;
 
     @GetMapping("/publish/{message}")
     public ResponseEntity<?> publishMessage(@PathVariable String message) {
         try {
-            for (int i = 0; i < 1_00_000; i++) {
-                publisher.sendMessageToTopic(i + " : " + message);
+            for (int i = 0; i < 10_000; i++) {
+                plainTextMessagePublisher.sendMessageToTopic(i + " : " + message);
             }
             return ResponseEntity.ok("Message published successfully ..");
         } catch (Exception exception) {
@@ -28,6 +32,6 @@ public class EventController {
 
     @PostMapping("/publish")
     public void sendEvents(@RequestBody Customer customer) {
-        publisher.sendEventsToTopic(customer);
+        eventMessagePublisher.sendEventsToTopic(customer);
     }
 }
