+++ kafka-producer-example/src/main/java/com/javatechie/service/KafkaPlainTextMessagePublisher.java
@@ -1,6 +1,5 @@
 package com.javatechie.service;
 
-import com.javatechie.dto.Customer;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.kafka.support.SendResult;
@@ -8,8 +7,10 @@ import org.springframework.stereotype.Service;
 
 import java.util.concurrent.CompletableFuture;
 
+import static com.javatechie.constants.KafkaTopicNames.TOPIC_NAME_TEXT_DATA;
+
 @Service
-public class KafkaMessagePublisher {
+public class KafkaPlainTextMessagePublisher {
 
     @Autowired
     private KafkaTemplate<String, Object> template;
@@ -21,7 +22,7 @@ public class KafkaMessagePublisher {
 
         // Use a topic which is already created
         // Behind the scene this topic is created with 3 partitions
-        CompletableFuture<SendResult<String, Object>> future = template.send("javatechie-demo-2", message);
+        CompletableFuture<SendResult<String, Object>> future = template.send(TOPIC_NAME_TEXT_DATA, message);
 
         future.whenComplete((result, ex) -> {
             if (ex == null) {
@@ -34,21 +35,4 @@ public class KafkaMessagePublisher {
         });
     }
 
-    public void sendEventsToTopic(Customer customer) {
-        try {
-            CompletableFuture<SendResult<String, Object>> future = template.send("javatechie-customer-1", customer);
-
-            future.whenComplete((result, ex) -> {
-                if (ex == null) {
-                    System.out.println("Sent customer = [" + customer.toString() +
-                            "] with offset=[" + result.getRecordMetadata().offset() + "]");
-                } else {
-                    System.out.println("Unable to send customer=[" +
-                            customer.toString() + "] due to : " + ex.getMessage());
-                }
-            });
-        } catch (Exception exception) {
-            System.out.println("EXCEPTION: " + exception.getMessage());
-        }
-    }
 }
